+++ greeting-service/src/main/resources/static/index.html
@@ -154,7 +154,7 @@
           <h1>Instructions</h1>
           <p>
               Please reference the mission documentation or <a
-                  href="https://github.com/snowdrop/spring-boot-istio-circuit-breaker-booster/blob/master/README.adoc">README</a>
+                  href="https://github.com/snowdrop/spring-boot-istio-circuit-breaker-example/blob/master/README.adoc">README</a>
               for next steps and instructions.
           </p>
       </div>
