+++ istio/initial_destination_rule.yml
@@ -1,7 +1,7 @@
 # We add the policy to disable TLS in order to ensure that
 # we can always use the same value of DISABLE in the
 # DestinationRule
-# This is only done to ensure that the booster works flawlessly in
+# This is only done to ensure that the example works flawlessly in
 # an Istio environment whether TLS is globally enabled or not
 apiVersion: authentication.istio.io/v1alpha1
 kind: Policy
