+++ name-service/src/main/java/dev/snowdrop/example/ExampleApplication.java
@@ -14,27 +14,19 @@
  * limitations under the License.
  */
 
-package io.openshift.booster.service;
+package dev.snowdrop.example;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 /**
- * Circuit Breaker state.
+ * Entry point to the application.
  */
-public class CircuitBreakerState {
+@SpringBootApplication
+public class ExampleApplication {
 
-    static final CircuitBreakerState OPEN = new CircuitBreakerState("open");
-    static final CircuitBreakerState CLOSED = new CircuitBreakerState("closed");
-
-    private String state;
-
-    public CircuitBreakerState() {
-    }
-
-    public CircuitBreakerState(String state) {
-        this.state = state;
-    }
-
-    public String getState() {
-        return state;
+    public static void main(String[] args) {
+        SpringApplication.run(ExampleApplication.class, args);
     }
 
 }
