+++ pom.xml
@@ -22,12 +22,12 @@
     <artifactId>booster-parent</artifactId>
     <version>23</version>
   </parent>
-  <groupId>io.openshift.booster</groupId>
-  <artifactId>spring-boot-istio-circuit-breaker-parent</artifactId>
+  <groupId>dev.snowdrop.example</groupId>
+  <artifactId>istio-circuit-breaker</artifactId>
   <version>2.1.3-3-SNAPSHOT</version>
   <packaging>pom</packaging>
-  <name>Spring Boot - Istio - Circuit Breaker Booster</name>
-  <description>Spring Boot - Istio - Circuit Breaker Booster</description>
+  <name>Spring Boot - Istio - Circuit Breaker Example</name>
+  <description>Spring Boot - Istio - Circuit Breaker Example</description>
   <properties>
     <arquillian.version>1.4.0.Final</arquillian.version>
     <arquillian-cube.version>1.18.2</arquillian-cube.version>
