+++ tests/pom.xml
@@ -4,14 +4,14 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <groupId>io.openshift.booster</groupId>
-    <artifactId>spring-boot-istio-circuit-breaker-parent</artifactId>
+    <groupId>dev.snowdrop.example</groupId>
+    <artifactId>istio-circuit-breaker</artifactId>
     <version>2.1.3-3-SNAPSHOT</version>
   </parent>
 
   <artifactId>spring-boot-istio-circuit-breaker-tests</artifactId>
 
-  <name>Spring Boot - Istio - Circuit Breaker Booster - Tests</name>
+  <name>Spring Boot - Istio - Circuit Breaker Example - Tests</name>
 
   <dependencies>
     <dependency>
@@ -120,12 +120,12 @@
             -->
             <dependencies>
               <dependency>
-                <groupId>io.openshift.booster</groupId>
+                <groupId>dev.snowdrop.example</groupId>
                 <artifactId>spring-boot-istio-circuit-breaker-name</artifactId>
                 <version>${project.version}</version>
               </dependency>
               <dependency>
-                <groupId>io.openshift.booster</groupId>
+                <groupId>dev.snowdrop.example</groupId>
                 <artifactId>spring-boot-istio-circuit-breaker-greeting</artifactId>
                 <version>${project.version}</version>
               </dependency>
