+++ tests/src/test/java/dev/snowdrop/example/GreetingWorker.java
@@ -1,4 +1,4 @@
-package io.openshift.booster;
+package dev.snowdrop.example;
 
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
